refactor: update tag colors
