+++ packages/wallet-management/src/components/EVMListItemButton.tsx
@@ -9,8 +9,8 @@ import { WalletManagementEvent } from '../types/events.js'
 import { createWalletConnectElement } from '../utils/elements.js'
 import { getConnectorIcon } from '../utils/getConnectorIcon.js'
 import { isWalletInstalled } from '../utils/isWalletInstalled.js'
+import { getTagType } from '../utils/walletTags.js'
 import { CardListItemButton } from './CardListItemButton.js'
-import { getTagType } from './WalletTag.js'
 import type { WalletListItemButtonProps } from './types.js'
 
 interface EVMListItemButtonProps extends WalletListItemButtonProps {
